Add test_star_probs and factor out relevant tests
