+++ chandra_aca/tests/test_all.py
@@ -112,37 +112,6 @@ def test_aca_targ_transforms():
     assert np.degrees(np.abs(dq.q[2] * 2)) < 1 / 3600.
 
 
-def test_t_ccd_warm_limit():
-    out = t_ccd_warm_limit([10.4] * 6, date='2015:001', min_n_acq=(2, 8e-3))
-    assert np.allclose(out[0], -15.2449, atol=0.01, rtol=0)
-    assert np.allclose(out[1], 0.008, atol=0.0001, rtol=0)
-
-    out = t_ccd_warm_limit([10.4] * 6, date='2015:001', min_n_acq=5.0)
-    assert np.allclose(out[0], -15.143, atol=0.01, rtol=0)
-    assert np.allclose(out[1], 5.0, atol=0.01, rtol=0)
-
-
-def test_mag_for_p_acq():
-    mag = mag_for_p_acq(0.50, date='2015:001', t_ccd=-14.0)
-    assert np.allclose(mag, 10.821, rtol=0, atol=0.01)
-
-
-def test_halfwidth_adjustment():
-    for mag, mults in ((6.0, [1.0, 1.0, 1.0, 1.0, 1.0]),
-                       (8.5, [1.0, 1.0, 1.0, 1.0, 1.0]),
-                       (9.25, [0.9, 0.9, 0.8, 0.7, 0.505]),
-                       (10.0, [0.8, 0.8, 0.6, 0.4, 0.01]),
-                       (11.0, [0.01, 0.01, 0.01, 0.01, 0.01])):
-        pacq = acq_success_prob(mag=mag, halfwidth=120)
-        p135 = acq_success_prob(mag=mag, halfwidth=135)
-        p160 = acq_success_prob(mag=mag, halfwidth=160)
-        p180 = acq_success_prob(mag=mag, halfwidth=180)
-        p240 = acq_success_prob(mag=mag, halfwidth=240)
-        p280 = acq_success_prob(mag=mag, halfwidth=280)
-        exp_mults = np.array([p135, p160, p180, p240, p280]) / pacq
-        assert np.allclose(mults, exp_mults)
-
-
 def test_get_aimpoint():
     obstests = [('2016-08-22', 15, 'ACIS-S'),
                ('2014-08-22', 16, 'HRC-I', True),
